backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
alembic==1.13.1
celery==5.3.4
redis==5.0.1
openai==1.12.0
requests==2.31.0